+++ src/controls.rs
@@ -5,10 +5,13 @@ use std::time::Duration;
 pub(super) struct Controls {
     /// Fade duration when play/pause
     pub(super) fade_duration: Duration,
-    /// When true, playback plays, when false playback is paused
-    pub(super) play: bool,
+    /// How long before source end should we send the prefetch notify callback.
+    /// Zero means don't send notify prefetch.
+    pub(super) prefetch: Duration,
     /// Sets the volume of the playback
     pub(super) volume: f32,
+    /// When true, playback plays, when false playback is paused
+    pub(super) play: bool,
 }
 
 impl Controls {
@@ -16,6 +19,7 @@ impl Controls {
     pub(super) fn new() -> Self {
         Self {
             fade_duration: Duration::ZERO,
+            prefetch: Duration::ZERO,
             play: false,
             volume: 1.,
         }